Fragment Implementation Deprecated
It was too hard to convert everything into fragments, I learned that fragments are a design pattern and should be __considered__ before start coding. Only two fragments worked but the communication between the main activity and the fragments sucked.

Luckily instead of showing a different fragment for each tab at the bottom bar instead is going to work as a 'button bar' which each press will be treated as a onClickListener.

 * Code will not be cleaned up just yet because __MAYBE__ I continue to implement fragments in a latter version.